pyproject.toml
[tool.poetry]
name = "commanderbot"
version = "0.16.0a1"
description = "A collection of utilities and extensions for discord.py bots."
authors = ["Arcensoth <arcensoth@gmail.com>"]
license = "MIT"
readme = 'README.md'
repository = "https://github.com/CommanderBot-Dev/commanderbot-py"
homepage = "https://github.com/CommanderBot-Dev/commanderbot-py"
keywords = ['discord', 'bot']

[tool.poetry.dependencies]
python = "^3.10"

# This is the last announced version of the discord.py 2.0 beta. RIP.
"discord.py" = { git = "https://github.com/Rapptz/discord.py.git", rev = "848d752" }

# These are used for logging.
colorama = "^0.4.3"
colorlog = "^4.2.1"

# NOTE See: https://github.com/python/typing/issues/865
# This exact version restriction can be removed once this bug is fixed.
typing-extensions = "3.10.0.0"

# The rest are for various extensions.

# NOTE This will be necessary until nbtlib updates numpy itself.
numpy = { version = "^1.21", python=">=3.10,<3.11" }
nbtlib = { git = "https://github.com/vberlier/nbtlib", branch = "main" }

# NOTE See: https://github.com/python-poetry/poetry/issues/4402
# This was fixed in stable, but doesn't seem to be fixed in the beta branch.
SQLAlchemy = ">=1.4,<1.4.23"

aiosqlite = "^0.17.0"
lectern = ">=0.15.0"
beet = ">=0.36.0"
aiohttp = "^3.7.4"
PyYAML = "^5.4.1"
emoji = "^1.4.2"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
isort = "^5.9.3"
rope = "^0.19.0"

[tool.isort]
profile = "black"

[tool.pyright]
typeCheckingMode = "basic"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
